refactor(config): :test_tube: move helpers class inside the folders helpers, remove useless imports
