+++ addons/gdUnit4/GdUnitRunner.cfg
@@ -9,7 +9,7 @@
 			"display_name": "InitializeMenus_Integration_FullWorkflowSuccess",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_Integration_FullWorkflowSuccess",
 			"guid": "4debf021-c89b-4140-8b7a-a607755b24fd",
-			"line_number": 176,
+			"line_number": 173,
 			"metadata": {
 
 			},
@@ -27,7 +27,7 @@
 			"display_name": "InitializeMenus_WithNullMenuContainer_DoesNotInitialize",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_WithNullMenuContainer_DoesNotInitialize",
 			"guid": "359571a5-00ff-4eb6-b439-5ba17211dbf0",
-			"line_number": 130,
+			"line_number": 127,
 			"metadata": {
 
 			},
@@ -45,7 +45,7 @@
 			"display_name": "InitializeMenus_WithNullMenuManager_DoesNotInitialize",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_WithNullMenuManager_DoesNotInitialize",
 			"guid": "0863cedf-55b6-49f3-8efd-9a8e319a699d",
-			"line_number": 75,
+			"line_number": 72,
 			"metadata": {
 
 			},
@@ -63,7 +63,7 @@
 			"display_name": "InitializeMenus_WithNullSettingsManager_DoesNotInitialize",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_WithNullSettingsManager_DoesNotInitialize",
 			"guid": "3d95fe80-96cb-4c10-907e-623d9277a4ee",
-			"line_number": 50,
+			"line_number": 47,
 			"metadata": {
 
 			},
@@ -81,7 +81,7 @@
 			"display_name": "InitializeMenus_WithValidInputs_InstantiatesAndAddsMenus",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_WithValidInputs_InstantiatesAndAddsMenus",
 			"guid": "3392c3ba-78a5-437e-809b-2bfcd27719f8",
-			"line_number": 100,
+			"line_number": 97,
 			"metadata": {
 
 			},
@@ -99,7 +99,7 @@
 			"display_name": "InitializeMenus_WithValidInputs_RegistersAndShowsMenus",
 			"fully_qualified_name": "UnitTests.MainTest.InitializeMenus_WithValidInputs_RegistersAndShowsMenus",
 			"guid": "bd2a915f-d2b7-4904-a885-2cfcc3eabfc7",
-			"line_number": 151,
+			"line_number": 148,
 			"metadata": {
 
 			},
