+++ scripts/Main.cs
@@ -1,5 +1,4 @@
 using AshesOfVelsingrad.Managers;
-using AshesOfVelsingrad.UI.Menus;
 using Godot;
 
 namespace AshesOfVelsingrad;
