+++ scripts/ui/menus/MainMenu.cs
@@ -1,4 +1,3 @@
-using System;
 using AshesOfVelsingrad.Managers;
 using Godot;
