+++ tests/helpers/TestSettingsManager.cs.uid
@@ -1,0 +1,1 @@
+uid://b4s70pb71src4
