+++ tests/helpers/TestableMain.cs
@@ -1,5 +1,3 @@
-using AshesOfVelsingrad.Managers;
-using AshesOfVelsingrad.UI.Menus;
 using Godot;
 using System.Collections.Generic;
